+++ stories/index.tsx
@@ -96,3 +96,5 @@ import '../src/topBar/story'
 import '../src/tripCard/story'
 import '../src/uneditableTextField/story'
 import '../src/why/story'
+// INTERNAL
+import '../src/_utils/rideAxis/story'