+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[NEW]** `RideAxis` component to display origin-destination
 - **[NEW]** Component `ItemBigData` to display a specific and important data.
 - **[UPDATE]** Use named exports instead of default exports.
 - **[FIX]** Fix Focus issue on `ItemChoice` using `FocusVisibleContext`.